requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiohttp==3.10.11
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
